test_requirements.txt
flask==1.0.2
flake8==3.7.7
pytest==4.5.0
requests-mock==1.6.0